Fix Hessian follow-ups from PR #177 review: ECP leak, 64-bit sizes, open-shell Raman
Five robustness/correctness follow-ups identified while validating the merged
analytic-Hessian PR (#177):

* ECP result-buffer leak: all four libecpint call sites (add_ecpint,
  add_ecpder, ecp_deriv_ints, add_ecphess) nulled result_ptr%data BEFORE
  calling free_result, which takes the struct by value -- so delete[] always
  received a null pointer and the C-side buffer leaked. Worst for the
  second-derivative payload (3N(3N+1)/2 * nbf^2 doubles per Hessian).
  free_result is now called first.

* 64-bit result sizes: ResultArray.size / ecp_result%size and the offset
  arithmetic (total_size, index, full_size, mat_sz) were 32-bit; the
  second-derivative element count overflows INT_MAX around ~50 atoms /
  500 basis functions. Now int64_t / integer(c_int64_t) end to end, with a
  comment documenting the remaining ~2x-payload peak-memory constraint of
  the concatenating libecpint API.

* Open-shell Raman activities: the vibrational-intensity path called the
  closed-shell cphf_static_polarizability for every reference, silently
  producing wrong Raman activities for UHF/ROHF Hessian runs. The validated
  per-spin (UHF) and single-MO (ROHF) polarizability tensors are extracted
  out of their selftests into cphf_uhf_static_polarizability /
  cphf_rohf_static_polarizability, and the C entry point now dispatches on
  scftype, so every caller gets the correct kernel with no Python changes.
  New regression test: RHF/UHF/ROHF dispatched tensors must coincide for a
  closed-shell molecule.

* ROHF ECP-centre map: the atom->ECP-centre match (1e-6 coordinate
  tolerance) had no completeness check; an unmapped centre would silently
  freeze the ECP while its atom is displaced. Now aborts with a clear
  message instead.

* CPHF preconditioner: the closed-shell and UHF solvers inverted the bare
  orbital-energy gap diagonal; exact occ-vir degeneracy gave Inf/NaN in the
  PCG. Both now clamp via 1/sign(max(|d|,1e-8),d), matching the ROHF solver.

+++ source/wrapper/libecpint_wrapper.cpp
@@ -9,12 +9,16 @@
 #include <iostream>
 #include <memory>
 #include <iomanip>
+#include <cstdint>
 
 extern "C" {
 
     typedef struct {
         double *data;
-        int size;
+        // 64-bit: the second-derivative payload is 3N(3N+1)/2 * nbf^2 doubles,
+        // which overflows a 32-bit count already around ~50 atoms / 500 basis
+        // functions. (Mirrored by integer(c_int64_t) in source/ecpint.F90.)
+        int64_t size;
     } ResultArray;
 
     // Initialize integrator and set Gaussian basis
@@ -60,18 +64,18 @@ extern "C" {
         factory->compute_first_derivs();
         std::vector<std::shared_ptr<std::vector<double>>> first_derivs = factory->get_first_derivs();
 
-        int total_size = 0;
+        int64_t total_size = 0;
         for (const auto& vec : first_derivs) {
-            total_size += vec->size();
+            total_size += static_cast<int64_t>(vec->size());
         }
 
         ResultArray result;
         result.size = total_size;
         result.data = new double[total_size];
-        int index = 0;
+        int64_t index = 0;
         for (const auto& vec : first_derivs) {
             std::copy(vec->begin(), vec->end(), result.data + index);
-            index += vec->size();
+            index += static_cast<int64_t>(vec->size());
         }
 
         return result;
@@ -83,18 +87,23 @@ extern "C" {
         factory->compute_second_derivs();
         std::vector<std::shared_ptr<std::vector<double>>> second_derivs = factory->get_second_derivs();
 
-        int total_size = 0;
+        // NOTE: libecpint already holds the full second-derivative set
+        // internally, and this wrapper copies it into one more concatenated
+        // buffer, so the peak footprint is ~2x the payload
+        // (3N(3N+1)/2 * nbf^2 doubles). A chunked per-component hand-off
+        // would halve that, but is constrained by the libecpint API.
+        int64_t total_size = 0;
         for (const auto& vec : second_derivs) {
-            total_size += vec->size();
+            total_size += static_cast<int64_t>(vec->size());
         }
 
         ResultArray result;
         result.size = total_size;
         result.data = new double[total_size];
-        int index = 0;
+        int64_t index = 0;
         for (const auto& vec : second_derivs) {
             std::copy(vec->begin(), vec->end(), result.data + index);
-            index += vec->size();
+            index += static_cast<int64_t>(vec->size());
         }
 
         return result;
